+++ test/integration/tests/haddock-options/Main.hs
@@ -5,8 +5,8 @@ main = do
     -- Fails to work because BAR is defined here and FOO in stack file
     stackErr ["haddock", "--haddock-arguments", "--optghc=-DBAR"]
     stack ["clean"]
-    -- Works just fine
-    stack ["haddock"]
+    -- Works just fine, test #3099 while at it.
+    stack ["haddock", "--no-haddock-hyperlink-source"]
     stack ["clean"]
     -- Fails to work because we have bad argument
     stackErr ["haddock", "--haddock-arguments", "--stack_it_badhaddockargument"]